resolve review comments
