+++ backend/pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+pythonpath = .
+asyncio_mode = auto
+testpaths = tests
