+++ backend/services/linkedin/audio/linkedin_audio_service.py
@@ -112,6 +112,7 @@ class LinkedInAudioService:
                 "estimated_duration_seconds": round(estimated_duration, 1),
                 "narration_preview": narration_text[:200],
                 "generation_time": generation_time,
+                "emotion": resolved_emotion,
             },
             user_id=user_id,
             file_extension=_LINKEDIN_AUDIO_FORMAT,
@@ -138,6 +139,7 @@ class LinkedInAudioService:
                 "target_duration_seconds": target_duration_seconds,
                 "generation_time": generation_time,
                 "format": _LINKEDIN_AUDIO_FORMAT,
+                "emotion": resolved_emotion,
                 "topic": topic,
                 "industry": industry,
             },
